♻️ refactor: 초기 데이터 로딩 부분을 커스텀 훅으로 분리
